Refactor dos testes de procedencia
